Moving config files to core/config
