+++ server/core/config/raw/policies.go
@@ -3,7 +3,7 @@ package raw
 import (
 	validation "github.com/go-ozzo/ozzo-validation"
 	"github.com/hashicorp/go-version"
-	"github.com/runatlantis/atlantis/server/events/yaml/valid"
+	"github.com/runatlantis/atlantis/server/core/config/valid"
 )
 
 // PolicySets is the raw schema for repo-level atlantis.yaml config.
