+++ server/core/config/raw/repo_cfg.go
@@ -4,7 +4,7 @@ import (
 	"errors"
 
 	validation "github.com/go-ozzo/ozzo-validation"
-	"github.com/runatlantis/atlantis/server/events/yaml/valid"
+	"github.com/runatlantis/atlantis/server/core/config/valid"
 )
 
 // DefaultAutomerge is the default setting for automerge.
