+++ server/core/config/raw/stage_test.go
@@ -4,8 +4,8 @@ import (
 	"testing"
 
 	validation "github.com/go-ozzo/ozzo-validation"
-	"github.com/runatlantis/atlantis/server/events/yaml/raw"
-	"github.com/runatlantis/atlantis/server/events/yaml/valid"
+	"github.com/runatlantis/atlantis/server/core/config/raw"
+	"github.com/runatlantis/atlantis/server/core/config/valid"
 	. "github.com/runatlantis/atlantis/testing"
 	yaml "gopkg.in/yaml.v2"
 )
