+++ server/core/config/valid/repo_cfg.go
