+++ server/events/models/models.go
@@ -28,7 +28,7 @@ import (
 	"github.com/runatlantis/atlantis/server/logging"
 
 	"github.com/pkg/errors"
-	"github.com/runatlantis/atlantis/server/events/yaml/valid"
+	"github.com/runatlantis/atlantis/server/core/config/valid"
 )
 
 const (
