+++ server/events/project_finder.go
@@ -19,7 +19,7 @@ import (
 	"path/filepath"
 	"strings"
 
-	"github.com/runatlantis/atlantis/server/events/yaml/valid"
+	"github.com/runatlantis/atlantis/server/core/config/valid"
 
 	"github.com/docker/docker/pkg/fileutils"
 	"github.com/pkg/errors"
